protect sql update from failing if triggers already exists (#38)
This helps for situations where the database schema version is downgraded to v17 by downgrading history-service. The database's schema doesn't actually change, but the service rewrites the version number from 18 to 17 anyway.

Unfortunately, this doesn't fix the underlying issue where the history-service can write schema version 17 to a v18 database.